+++ src/MicroElements.Metadata/Metadata/ParserRuleExtensions.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using MicroElements.Metadata.Parsers;
@@ -35,10 +36,15 @@ namespace MicroElements.Metadata
 
             if (parserRule == null)
             {
-                // Enum support.
                 if (property.Type.IsEnum)
                 {
+                    // Enum.
                     parserRule = new ParserRule(new EnumUntypedParser(property.Type), property.Type);
+                }
+                else if (Nullable.GetUnderlyingType(property.Type) is { } baseType && baseType.IsEnum)
+                {
+                    // Nullable enum.
+                    parserRule = new ParserRule(new EnumUntypedParser(baseType, allowNull: true), property.Type);
                 }
             }