+++ src/MicroElements.Metadata/Metadata/PropertyParserExtensions.cs
@@ -17,7 +17,7 @@ namespace MicroElements.Metadata
         /// <param name="propertyParser">Source property parser.</param>
         /// <param name="sourceRow">Value to parse.</param>
         /// <returns>Optional parse result.</returns>
-        public static Option<IPropertyValue> ParseRowUntyped(this IPropertyParser propertyParser, IReadOnlyDictionary<string, string> sourceRow)
+        public static ParseResult<IPropertyValue> ParseRowUntyped(this IPropertyParser propertyParser, IReadOnlyDictionary<string, string> sourceRow)
         {
             return sourceRow
                 .GetValueAsOption(propertyParser.SourceName)
@@ -30,12 +30,12 @@ namespace MicroElements.Metadata
         /// <param name="propertyParser">Source property parser.</param>
         /// <param name="textValue">Value to parse.</param>
         /// <returns>Optional parse result.</returns>
-        public static Option<IPropertyValue> ParseUntyped(this IPropertyParser propertyParser, string textValue)
+        public static ParseResult<IPropertyValue> ParseUntyped(this IPropertyParser propertyParser, string textValue)
         {
-            var func = CodeCompiler.CachedCompiledFunc<IPropertyParser, string, Option<IPropertyValue>>(propertyParser.TargetType, "Parse", Parse<CodeCompiler.GenericType>);
+            var func = CodeCompiler.CachedCompiledFunc<IPropertyParser, string, ParseResult<IPropertyValue>>(propertyParser.TargetType, "Parse", Parse<CodeCompiler.GenericType>);
             return func(propertyParser, textValue);
 
-            static Option<IPropertyValue> Parse<T>(IPropertyParser propertyParserUntyped, string textValue)
+            static ParseResult<IPropertyValue> Parse<T>(IPropertyParser propertyParserUntyped, string textValue)
             {
                 var propertyParserTyped = (IPropertyParser<T>)propertyParserUntyped;
 
@@ -51,21 +51,21 @@ namespace MicroElements.Metadata
         /// </summary>
         /// <param name="propertyParser">Source property parser.</param>
         /// <returns>Returns optional <see cref="IPropertyValue"/> if default value evaluated for <paramref name="propertyParser"/>.</returns>
-        public static Option<IPropertyValue> GetDefaultValueUntyped(this IPropertyParser propertyParser)
+        public static ParseResult<IPropertyValue> GetDefaultValueUntyped(this IPropertyParser propertyParser)
         {
-            var func = CodeCompiler.CachedCompiledFunc<IPropertyParser, Option<IPropertyValue>>(propertyParser.TargetType, "GetDefaultValue", GetDefaultValue<CodeCompiler.GenericType>);
+            var func = CodeCompiler.CachedCompiledFunc<IPropertyParser, ParseResult<IPropertyValue>>(propertyParser.TargetType, "GetDefaultValue", GetDefaultValue<CodeCompiler.GenericType>);
             return func(propertyParser);
 
-            static Option<IPropertyValue> GetDefaultValue<T>(IPropertyParser propertyParserUntyped)
+            static ParseResult<IPropertyValue> GetDefaultValue<T>(IPropertyParser propertyParserUntyped)
             {
                 IPropertyParser<T> propertyParser = (IPropertyParser<T>)propertyParserUntyped;
                 if (propertyParser.DefaultValue != null)
                 {
                     var defaultValue = propertyParser.DefaultValue();
-                    return new PropertyValue<T>(propertyParser.TargetProperty, defaultValue, ValueSource.DefaultValue);
+                    return ParseResult.Success<IPropertyValue>(new PropertyValue<T>(propertyParser.TargetProperty, defaultValue, ValueSource.DefaultValue));
                 }
 
-                return Option<IPropertyValue>.None;
+                return ParseResult<IPropertyValue>.Failed;
             }
         }
     }